+++ Keel/src/keel/facade/translators/__init__.py
@@ -1,4 +1,2 @@
 from keel.facade.translators.domain_to_facade import (
     domain_to_facade_run_report, domain_to_facade_step_report)
-from keel.facade.translators.facade_to_domain import \
-    facade_to_domain_run_request